Fix compilation error: cast literal to uint16_t for std::min type matching
Fixed type mismatch in temperature calculation where std::min(48, oa_raw)
failed because 48 is int and oa_raw is uint16_t. Cast 48 to uint16_t
explicitly.

Also answering question: The simulated app recognized the ESP32 as a
"MEATER Block" device (masterType=MASTER_TYPE_BLOCK), not a MEATER or
MEATER+ probe. The Block is the WiFi-enabled base station that connects
to probes via BLE and broadcasts their data over UDP to the app.

+++ includes/meater_udp_broadcast.h
@@ -320,7 +320,7 @@ class MeaterUDPBroadcaster {
       
       // Calculate temperatures in Celsius * 16 (as per protocol)
       int32_t tip_temp = tip_raw + 8;  // Raw value already includes offset
-      int32_t ambient_temp = tip_raw + std::max(0, (((ra_raw - std::min(48, oa_raw)) * 16 * 589)) / 1487) + 8;
+      int32_t ambient_temp = tip_raw + std::max(0, (((ra_raw - std::min((uint16_t)48, oa_raw)) * 16 * 589)) / 1487) + 8;
       
       encode_sint32_field(out, 1, tip_temp);        // internalTemperature
       encode_sint32_field(out, 2, ambient_temp);    // ambientTemperature
